Finished Worksheet 2

+++ P02/Converter.java
@@ -1,3 +1,8 @@
+/* *
+ * Luke McDougall 17179492
+ * Simple program that converts a user input temperature from celsius to fahrenheit
+ * Last updated 10/03/2020
+ * */
 import java.util.*;
 
 public class Converter
@@ -5,13 +10,13 @@ public class Converter
     public static void main(String[] args)
     {
         Scanner input = new Scanner(System.in);
-        int celsius;
+        double celsius;
         double fahrenheit;
 
         System.out.print("Please enter the temperature in Celsius: ");
-        celsius = input.nextInt();
-        fahrenheit = (9.0 / 5.0) * (double) celsius + 32.0;
+        celsius = input.nextDouble();
+        fahrenheit = (9.0 / 5.0) * celsius + 32.0;
 
-        System.out.printf("%d degrees Celsius = %.2f degrees Fahrenheit\n", celsius, fahrenheit);
+        System.out.printf("%.2f degrees Celsius = %.2f degrees Fahrenheit\n", celsius, fahrenheit);
     }
 }
